Consolidated files for Simplicity
