--- backend/Models/SupabaseSettings.cs
@@ -1,8 +1,0 @@
-namespace didaktos.backend.Models
-{
-    public class SupabaseSettings
-    {
-        public string Url { get; set; } = string.Empty;
-        public string Key { get; set; } = string.Empty;
-    }
-}
