--- backend/Repositories/IUserRepository.cs
@@ -1,11 +1,0 @@
-using didaktos.backend.Models;
-
-namespace didaktos.backend.Services
-{
-    public interface IUserRepository
-    {
-        Task<User?> GetByEmailAsync(string email);
-        Task<User?> GetByIdAsync(Guid id);
-        Task<User> CreateAsync(User user);
-    }
-}
