--- backend/Services/IAuthService.cs
@@ -1,10 +1,0 @@
-using didaktos.backend.Models.DTOs;
-
-namespace didaktos.backend.Services
-{
-    public interface IAuthService
-    {
-        Task<AuthResponseDto> RegisterAsync(RegisterRequestDto request);
-        Task<AuthResponseDto> LoginAsync(LoginRequestDto request);
-    }
-}
