--- backend/Services/IJwtTokenService.cs
@@ -1,9 +1,0 @@
-using didaktos.backend.Models;
-
-namespace didaktos.backend.Services
-{
-    public interface IJwtTokenService
-    {
-        string GenerateToken(User user);
-    }
-}
